Move linking config to NavigationConfig
