+++ src/App.tsx
@@ -18,50 +18,12 @@ import "react-native-gesture-handler";
 import { enableScreens } from "react-native-screens";
 import { useSelector } from "react-redux";
 import { StoreState } from "./store";
-import { RootStackParamList } from "./NavigationConfig";
+import { linking } from "./NavigationConfig";
 import { DarkTheme, LightTheme } from "./theme";
 
 enableScreens();
 
 const DrawerNavigation = createDrawerNavigator();
-
-type RootStackScreens = {
-  [route in keyof RootStackParamList]: string;
-};
-
-const rootStackScreens: RootStackScreens = {
-  Home: "",
-  Login: "login",
-  Signup: "signup",
-  CollectionCreate: "new-notebook",
-  Collection: "notebook/:colUid",
-  CollectionEdit: "notebook/:colUid/edit",
-  CollectionChangelog: "notebook/:colUid/manage",
-  CollectionMembers: "notebook/:colUid/members",
-  NoteCreate: "new-note",
-  NoteEdit: "notebook/:colUid/note/:itemUid",
-  NoteProps: "notebook/:colUid/note/:itemUid/properties",
-  NoteMove: "notebook/:colUid/note/:itemUid/move",
-  Invitations: "invitations",
-  Settings: "settings",
-  About: "settings/about",
-  Password: "settings/password",
-  DebugLogs: "settings/logs",
-  AccountWizard: "account-wizard",
-  "404": "*",
-};
-
-const linking = {
-  prefixes: ["https://notes.etesync.com", "com.etesync.notes://"],
-  config: {
-    screens: {
-      Root: {
-        path: "",
-        screens: rootStackScreens,
-      },
-    },
-  },
-};
 
 function InnerApp() {
   // XXX Workaround for react-navigation #7561 (flashing drawer)
